requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-telegram-bot==20.3
python-dotenv==1.0.0
apscheduler==3.10.4
requests==2.31.0
paramiko==3.4.0
psutil==5.9.6
aiohttp==3.9.1
httpx==0.24.1